Split up logic between main script, mapillary script, and new util script
