+++ mapillary.py
@@ -1,6 +1,10 @@
 import time
 from dateutil import parser
 import requests
+import os
+import pickle
+import multiprocessing
+import util
 
 SEQUENCES_PER_PAGE_DEFAULT = 50  # How many sequences to receive on each page of the API call
 SKIP_IF_FEWER_IMAGES_THAN_DEFAULT = 10  # We will skip any sequences if they have fewer than this number of images
@@ -8,14 +12,90 @@ SEQUENCE_URL = 'https://a.mapillary.com/v3/sequences_without_images?client_id={}
 IMAGES_URL = 'https://a.mapillary.com/v3/images?client_id={}&sequence_keys={}'
 
 
+def run(bbox: str, traces_source: any, processes: int):
+    # Do a quick check to see if user specified the mandatory 'client_id' in traces_source JSON
+    if 'client_id' not in traces_source:
+        raise KeyError('Missing "client_id" (Mapillary Client ID) key in --traces-source JSON.')
+
+    # Create dirs
+    output_dir, output_tmp_dir = util.initialize_dirs(bbox)
+
+    # Break the bbox into sections and save it to a pickle file
+    bbox_sections = util.split_bbox(output_dir, bbox, to_bbox)
+
+    finished_bbox_sections = multiprocessing.Value('i', 0)
+    with multiprocessing.Pool(initializer=initialize_multiprocess,
+                              initargs=(output_dir, output_tmp_dir, traces_source, finished_bbox_sections),
+                              processes=processes) as pool:
+        result = pool.map_async(pull_and_save_trace_for_bbox, bbox_sections)
+
+        print('Placing {} results in {}...'.format(len(bbox_sections), output_dir))
+        progress = 0
+        increment = 5
+        while not result.ready():
+            result.wait(timeout=5)
+            next_progress = int(finished_bbox_sections.value / len(bbox_sections) * 100)
+            if int(next_progress / increment) > progress:
+                print('Current progress: {}%'.format(next_progress))
+                progress = int(next_progress / increment)
+        if progress != 100 / increment:
+            print('Current progress: 100%')
+
+        # TODO: Delete the tmp dir after run?
+
+
 def to_bbox(llo, lla, mlo, mla):
     return ','.join([str(llo), str(lla), str(mlo), str(mla)])
 
 
-def get_trace_data_for_bbox(session: requests.Session, bbox: str, conf: any) -> dict[
-    str, list]:
+def initialize_multiprocess(global_output_dir_: str, global_output_tmp_dir_: str, global_traces_source_: any,
+                            finished_bbox_sections_: multiprocessing.Value):
+    # For persistent connections
+    global session
+    session = requests.Session()
+
+    # So each process knows the output / tmp dirs
+    global global_output_dir
+    global_output_dir = global_output_dir_
+    global global_output_tmp_dir
+    global_output_tmp_dir = global_output_tmp_dir_
+
+    # So each process knows the conf provided
+    global global_traces_source
+    global_traces_source = global_traces_source_
+
+    global finished_bbox_sections
+    finished_bbox_sections = finished_bbox_sections_
+
+
+def pull_and_save_trace_for_bbox(bbox: str) -> None:
+    try:
+        # The file on disk where we will store trace data
+        result_filename = os.path.join(global_output_dir, bbox + '.pickle')
+
+        if os.path.exists(result_filename):
+            print('Seq for bbox={} already exists on disk! Skipping...'.format(bbox))
+            with finished_bbox_sections.get_lock():
+                finished_bbox_sections.value += 1
+            return
+
+        # We haven't pulled API trace data for this bbox section yet
+        trace_data = trace_data_request(session, bbox, global_traces_source)
+
+        # Avoids potential partial write issues by writing to a temp file and then as a final operation, then renaming
+        # to the real location
+        temp_filename = os.path.join(global_output_tmp_dir, bbox + '.pickle')
+        pickle.dump(trace_data, open(temp_filename, 'wb'))
+        os.rename(temp_filename, result_filename)
+
+        with finished_bbox_sections.get_lock():
+            finished_bbox_sections.value += 1
+    except Exception as e:
+        print('ERROR: Failed to pull trace data: {}'.format(repr(e)))
+
+
+def trace_data_request(session_: requests.Session, bbox: str, conf: any) -> dict[str, list]:
     result, elapsed = {}, 0
-    # try:
     start = time.time()
 
     map_client_id = conf['client_id']  # The Mapillary client ID, mandatory key of conf
@@ -30,14 +110,14 @@ def get_trace_data_for_bbox(session: requests.Session, bbox: str, conf: any) -> 
     page = 1
     while next_url:
         print('Page {}, url={}'.format(page, next_url))
-        sequence_resp = session.get(next_url)
+        sequence_resp = session_.get(next_url)
         sequence_keys = []
         for seq_f in sequence_resp.json()['features']:
             # Skip sequences that have too few images
             if len(seq_f['geometry']['coordinates']) >= skip_if_fewer_imgs_than:
                 sequence_keys.append(seq_f['properties']['key'])
         # TODO: Paginate images
-        images_resp = session.get(IMAGES_URL.format(map_client_id, ','.join(sequence_keys)))
+        images_resp = session_.get(IMAGES_URL.format(map_client_id, ','.join(sequence_keys)))
         for img_f in images_resp.json()['features']:
             if img_f['properties']['sequence_key'] not in result:
                 result[img_f['properties']['sequence_key']] = []
@@ -54,9 +134,5 @@ def get_trace_data_for_bbox(session: requests.Session, bbox: str, conf: any) -> 
     elapsed = stop - start
     print('\n\n##################\nFinished processing seqs for bbox={}, elapsed time: {}'.format(bbox, elapsed))
     print('{}\n##################\n\n'.format(result))
-    # with response_count.get_lock():
-    #     response_count.value += 1
-    # except Exception as e:
-    #     print(e)
 
     return result
